Move row lib into cli crate.
