+++ src/cli/autocomplete.rs
@@ -4,9 +4,9 @@
 use clap_complete::CompletionCandidate;
 use indicatif::{MultiProgress, ProgressDrawTarget};
 
-use row::MultiProgressContainer;
-use row::workflow::Workflow;
-use row::{cluster, workspace};
+use crate::MultiProgressContainer;
+use crate::workflow::Workflow;
+use crate::{cluster, workspace};
 
 /// List the actions in the current workflow.
 pub fn get_action_candidates() -> Vec<CompletionCandidate> {
