+++ src/cli/clean.rs
@@ -7,9 +7,9 @@ use std::error::Error;
 use std::{fs, io};
 
 use crate::cli::GlobalOptions;
-use row::MultiProgressContainer;
-use row::project::Project;
-use row::{
+use crate::MultiProgressContainer;
+use crate::project::Project;
+use crate::{
     COMPLETED_CACHE_FILE_NAME, DATA_DIRECTORY_NAME, DIRECTORY_CACHE_FILE_NAME,
     SUBMITTED_CACHE_FILE_NAME,
 };
@@ -73,7 +73,7 @@ pub fn clean(
         }
         if force_needed && !args.force {
             warn!("You should wait for these jobs to complete.");
-            return Err(Box::new(row::Error::ForceCleanNeeded));
+            return Err(Box::new(crate::Error::ForceCleanNeeded));
         }
     }
 
@@ -85,7 +85,7 @@ pub fn clean(
         if let Err(error) = fs::remove_file(&path) {
             match error.kind() {
                 io::ErrorKind::NotFound => (),
-                _ => return Err(Box::new(row::Error::FileRemove(path.clone(), error))),
+                _ => return Err(Box::new(crate::Error::FileRemove(path.clone(), error))),
             }
         }
     }
@@ -95,7 +95,7 @@ pub fn clean(
         if let Err(error) = fs::remove_file(&path) {
             match error.kind() {
                 io::ErrorKind::NotFound => (),
-                _ => return Err(Box::new(row::Error::FileRemove(path.clone(), error))),
+                _ => return Err(Box::new(crate::Error::FileRemove(path.clone(), error))),
             }
         }
     }
@@ -105,7 +105,7 @@ pub fn clean(
         if let Err(error) = fs::remove_file(&path) {
             match error.kind() {
                 io::ErrorKind::NotFound => (),
-                _ => return Err(Box::new(row::Error::FileRemove(path.clone(), error))),
+                _ => return Err(Box::new(crate::Error::FileRemove(path.clone(), error))),
             }
         }
     }
