+++ src/cli/directories.rs
@@ -12,8 +12,8 @@ use std::path::PathBuf;
 
 use crate::cli::{self, GlobalOptions, autocomplete};
 use crate::ui::{Alignment, Item, Row, Table};
-use row::MultiProgressContainer;
-use row::project::Project;
+use crate::MultiProgressContainer;
+use crate::project::Project;
 
 #[derive(Args, Debug)]
 pub struct Arguments {
@@ -112,7 +112,7 @@ pub fn print_matching<W: Write>(
     project
         .workflow()
         .action_by_name(action_name)
-        .ok_or_else(|| row::Error::ActionNotFound(action_name.to_string()))?;
+        .ok_or_else(|| crate::Error::ActionNotFound(action_name.to_string()))?;
 
     let mut table = Table::new().with_hide_header(if args.short { true } else { args.no_header });
     table.header = vec![
@@ -222,7 +222,7 @@ pub fn print_matching<W: Write>(
                     let value = project.state().values()[directory]
                         .pointer(pointer)
                         .ok_or_else(|| {
-                            row::Error::JSONPointerNotFound(directory.clone(), pointer.clone())
+                            crate::Error::JSONPointerNotFound(directory.clone(), pointer.clone())
                         })?;
                     row.push(
                         Item::new(value.to_string(), Style::new()).with_alignment(Alignment::Right),
