+++ src/cli/jobs.rs
@@ -13,8 +13,8 @@ use wildmatch::WildMatch;
 
 use crate::cli::{self, GlobalOptions, autocomplete};
 use crate::ui::{Alignment, Item, Row, Table};
-use row::MultiProgressContainer;
-use row::project::Project;
+use crate::MultiProgressContainer;
+use crate::project::Project;
 
 #[derive(Args, Debug)]
 pub struct Arguments {
