+++ src/cli/status.rs
@@ -13,9 +13,9 @@ use wildmatch::WildMatch;
 
 use crate::cli::{self, GlobalOptions, autocomplete};
 use crate::ui::{Alignment, Item, Row, Table};
-use row::MultiProgressContainer;
-use row::project::{Project, Status};
-use row::workflow::ResourceCost;
+use crate::MultiProgressContainer;
+use crate::project::{Project, Status};
+use crate::workflow::ResourceCost;
 
 #[derive(Args, Debug)]
 pub struct Arguments {
