+++ src/cli/submit.rs
@@ -20,10 +20,10 @@ use std::time::Instant;
 use wildmatch::WildMatch;
 
 use crate::cli::{GlobalOptions, autocomplete};
-use row::MultiProgressContainer;
-use row::format::HumanDuration;
-use row::project::Project;
-use row::workflow::{Action, ResourceCost};
+use crate::MultiProgressContainer;
+use crate::format::HumanDuration;
+use crate::project::Project;
+use crate::workflow::{Action, ResourceCost};
 
 #[derive(Args, Debug)]
 pub struct Arguments {
@@ -97,7 +97,7 @@ pub fn submit<W: Write>(
             )?;
             for group in &groups {
                 if !whole_groups.contains(group) {
-                    return Err(Box::new(row::Error::PartialGroupSubmission(
+                    return Err(Box::new(crate::Error::PartialGroupSubmission(
                         action.name().into(),
                     )));
                 }
@@ -107,7 +107,7 @@ pub fn submit<W: Write>(
         for group in &groups {
             for directory in group {
                 if !action_directory_set.insert((action.name.clone(), directory.clone())) {
-                    return Err(Box::new(row::Error::WouldSubmitMultipleTimes(
+                    return Err(Box::new(crate::Error::WouldSubmitMultipleTimes(
                         directory.clone(),
                         action.name().into(),
                     )));
